+++ configs/all.js
@@ -123,3 +123,6 @@ module.exports = {
   },
   rules: allRules,
 };
+
+// this is so the `languageOptions` property won't be warned in the new config system
+Object.defineProperty(module.exports, 'languageOptions', { enumerable: false });